A generator for more-highly checked boards
Change the wordlist and modify the function to determine if two words are close together in #configurable settings. Then you should just be able to run rand_board() / rand_board(minlen,toplen,height,width) to generate a board as necessary. rand_board returns a board object, use repr(board) to get a string output (cropped to the size of the board) or board.output() to get a list of word placements (uncropped).